pyproject.toml
[project]
name = "omsktsifra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "authx>=1.4.2",
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.13.4",
    "database",
    "databases==0.9.0",
    "fastapi>=0.115.12",
    "lxml>=6.0.0",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pyright>=1.1.402",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
